Product projections search integration tests

+++ commercetools.Sdk/commercetools.Sdk.Domain/Query/FacetTerm.cs
@@ -3,7 +3,7 @@
     public class FacetTerm
     {
         public string Term { get; set; }
-        public double Count { get; set; }
-        public double ProductCount { get; set; }
+        public long Count { get; set; }
+        public long ProductCount { get; set; }
     }
 }
